Original program file on Google colab with output
Note: color choice is random except for 4 students for which I wanted a particular color to fit the text. With this program color could be chosen twice in a row because of this. We can see that it happened with Catherine and Audrey presentations. It didn't matter as the text between introduction texts is white and because it  was just a little program written for fun ;-).